requirements.txt
groq
python-dotenv
Flask
uuid
gtts